Iterator: use get_unchecked(), makes runtime 2x faster
This is in the most extreme cases (N >= 1M).

Note that it is OK to use get_unchecked() in this case,
as we have an invariat that has to be kept - the size of
all vectors has to be the same at all times.

This means that if we are in the bounds of one vector,
we are in the bounds of all vectors.

This way we can skip this check being made for every
field in the struct.